Add test for song
Since tests should be concrete we avoid all abstractions and dependencies so that the test doesn't rely on the implementation details. That means we literally need the whole song in a string somewhere. I put it in a txt file and hardcoded the path for the time being. I'll handle this better later.