Remove the old method calls by caching the BottleNumber in a private var
We can greatly reduce the number of objects created by caching the current bottle number in a local var. This is very cheap and easy to understand so it's a good trade. Also it's come to light that Successor method should return a new bottle number. We used a hack of creating a second bottle number for now.