Override ToString so it returns Quantity and Container on bottle number
This removes a small data clump where we repeatedly called the same combo of methods in order over and over. By moving those two things to ToString and taking advantage of the fact String interpolation automatically calls ToString on the result it gets from the evaluation we can greatly simplify verse. The only exception is where we're using Capitalize which requires a string input.